requirements.txt
langgraph
langchain
langchain-community
langchain-google-genai
langchain-openai
duckduckgo-search>=6.1.0
python-dotenv

## always ensure compatibility, some not mentioned which versions to use as of now